chore(linux): drop stale comments, dedupe verification-script measurement
- app/linux-loopback.js: header no longer describes the module as
  unwired/spike-only — it's fully integrated (main.js, IPC, renderer).
- build-linux.yml: removed the GPU-pruning "known gap" paragraph — verified
  false; stenoai.spec's GPU-marker skip is unconditional, not Windows-only.
- scripts/measure-pcm.js: shared peak/RMS helper, replacing the same loop
  hand-rolled in both linux-loopback-poc-demo.js and
  app/scripts/linux-loopback-e2e-check.js.
- linux-loopback-poc-demo.js: drop the redundant getDefaultSinkName() call
  now that startLoopbackCapture() resolves it internally and reports it
  back via capture.target.

Re-verified both scripts against the real PipeWire/Electron app after the
change — same real-audio pass as before.
